+++ Src/01/06/KSociety.Base.Infra.Shared/Class/UnitOfWork.cs
@@ -27,10 +27,10 @@ namespace KSociety.Base.Infra.Shared.Class
         }
 
         ///<inheritdoc/>
-        public ValueTask<string> GetConnectionStringAsync(CancellationToken cancellationToken = default)
+        public ValueTask<string?> GetConnectionStringAsync(CancellationToken cancellationToken = default)
         {
             var result = Context?.GetConnectionString();
-            return new ValueTask<string>(result);
+            return new ValueTask<string?>(result);
         }
 
         ///<inheritdoc/>
@@ -48,7 +48,8 @@ namespace KSociety.Base.Infra.Shared.Class
         ///<inheritdoc/>
         public async ValueTask<bool?> EnsureCreatedAsync(CancellationToken cancellationToken = default)
         {
-            return await Context.EnsureCreatedAsync(cancellationToken).ConfigureAwait(false);
+            if (Context != null) return await Context.EnsureCreatedAsync(cancellationToken).ConfigureAwait(false);
+            return null;
         }
 
         ///<inheritdoc/>
@@ -60,7 +61,8 @@ namespace KSociety.Base.Infra.Shared.Class
         ///<inheritdoc/>
         public async ValueTask<bool?> EnsureDeletedAsync(CancellationToken cancellationToken = default)
         {
-            return await Context.EnsureDeletedAsync(cancellationToken).ConfigureAwait(false);
+            if (Context != null) return await Context.EnsureDeletedAsync(cancellationToken).ConfigureAwait(false);
+            return null;
         }
 
         ///<inheritdoc/>
@@ -73,7 +75,7 @@ namespace KSociety.Base.Infra.Shared.Class
         public async ValueTask MigrateAsync(string? targetMigration = null,
             CancellationToken cancellationToken = default)
         {
-            await Context.MigrateAsync(targetMigration, cancellationToken).ConfigureAwait(false);
+            if (Context != null) await Context.MigrateAsync(targetMigration, cancellationToken).ConfigureAwait(false);
         }
 
         public string? CreateScript()
@@ -90,7 +92,7 @@ namespace KSociety.Base.Infra.Shared.Class
         ///<inheritdoc/>
         public async ValueTask BeginTransactionAsync(CancellationToken cancellationToken = default)
         {
-            await Context.BeginTransactionAsync(cancellationToken).ConfigureAwait(false);
+            if (Context != null) await Context.BeginTransactionAsync(cancellationToken).ConfigureAwait(false);
         }
 
         ///<inheritdoc/>
@@ -102,7 +104,8 @@ namespace KSociety.Base.Infra.Shared.Class
         ///<inheritdoc/>
         public async ValueTask<int?> CommitAsync(CancellationToken cancellationToken = default)
         {
-            return await Context.CommitAsync(cancellationToken).ConfigureAwait(false);
+            if (Context != null) return await Context.CommitAsync(cancellationToken).ConfigureAwait(false);
+            return null;
         }
 
         ///<inheritdoc/>
@@ -114,7 +117,7 @@ namespace KSociety.Base.Infra.Shared.Class
         ///<inheritdoc/>
         public async ValueTask CommitTransactionAsync(CancellationToken cancellationToken = default)
         {
-            await Context.CommitTransactionAsync(cancellationToken).ConfigureAwait(false);
+            if (Context != null) await Context.CommitTransactionAsync(cancellationToken).ConfigureAwait(false);
         }
 
         ///<inheritdoc/>
@@ -126,7 +129,7 @@ namespace KSociety.Base.Infra.Shared.Class
         ///<inheritdoc/>
         public async ValueTask RollbackAsync(CancellationToken cancellationToken = default)
         {
-            await Context.RollbackAsync(cancellationToken).ConfigureAwait(false);
+            if(Context != null) await Context.RollbackAsync(cancellationToken).ConfigureAwait(false);
         }
 
         ///<inheritdoc/>
